ch 10 traits trait is smilar to interface in other programming language, but they also have difference
we can implement a trait on a type
trait methods can have default impelmentation
trait can also used as parameter to define function. This parameter accpets any value type that implements the trait
we can add trait bounds to limit the parameter type
trait can also be used in return types to indicate the return value type implements a trait, but we can't return multiple value types that implements the same trait
we can use trait bounds to conditionaly impelments methods

We can also conditionally implement a trait for any type that implements another trait. Implementations of a trait on any type that satisfies the trait bounds are called blanket implementations and are extensively used in the Rust standard library.