13. Functional Language Features: Iterators and Closures Rust’s closures are anonymous functions you can save in a variable or pass as arguments to other functions. Closures can capture values from the scope in which they’re defined. Closures don’t require you to annotate the types of the parameters or the return value like fn functions do. The compiler can usually infer the parameter and return types Closures are usually short and relevant only within a narrow context rather than in any arbitrary scenario.
Closures can capture values from their environment in three ways, which directly map to the three ways a function can take a parameter: taking ownership, borrowing mutably, and borrowing immutably. These are encoded in the three Fn traits as follows:

* FnOnce consumes the variables it captures from its enclosing scope, known as the closure’s environment. To consume the captured variables, the closure must take ownership of these variables and move them into the closure when it is defined. The Once part of the name represents the fact that the closure can’t take ownership of the same variables more than once, so it can be called only once.
* FnMut can change the environment because it mutably borrows values.
* Fn borrows values from the environment immutably.

We can use `move` keyword before closure parameter list to take ownership of **captured values** from environment. This technique is mostly useful when passing a closure to a new thread to move the data so it’s owned by the new thread.