15.3. Running Code on Cleanup with the Drop Trait
we can add custom behavior to value when they are out of scope by implements Drop trait. They can be used to release resources like files or network connections. and the drop logic is inserted by compiler whenever the value is out of scope

Variables are dropped in the reverse order of their creation

we can also drop a value early by call std::mem::drop.

We can't call drop method on the value explicitly