15.4. Rc, the Reference Counted Smart Pointer
To enable multiple ownership, Rust has a type called Rc<T>, which is an abbreviation for reference counting. The Rc<T> type keeps track of the number of references to a value which determines whether or not a value is still in use. If there are zero references to a value, the value can be cleaned up without any references becoming invalid.

Via immutable references, Rc<T> allows you to share data between multiple parts of your program for reading only.

The call to Rc::clone only increments the reference count, which doesn’t take much time.

call Rc::clone to increase the reference count

We don’t have to call a function to decrease the reference count . The implementation of the Drop trait decreases the reference count automatically when an Rc<T> value goes out of scope.