17. Object Oriented Programming Features of Rust
17.1. Characteristics of Object-Oriented Languages

Object-oriented programs are made up of objects. An object packages both data and the procedures that operate on that data. The procedures are typically called methods or operations.

`object`: structs and enums have data, and impl blocks provide methods on structs and enums.

encapsulation: use the pub keyword to decide which modules, types, functions, and methods in our code should be public, and by default everything else is private.

Inheritance
You choose inheritance for two main reasons.
* One is for reuse of code: you can implement particular behavior for one type, and inheritance enables you to reuse that implementation for a different type. You can share Rust code using default trait method implementations instead
* The other reason to use inheritance relates to the type system: to enable a child type to be used in the same places as the parent type. This is also called polymorphism, which means that you can substitute multiple objects for each other at runtime if they share certain characteristics. In rust we uses generics to abstract over different possible types and trait bounds to impose constraints on what those types must provide. This is sometimes called bounded parametric polymorphism.