18.3. Pattern Syntax
Matching Literals
Matching Named Variables
Multiple Patterns

Destructuring to Break Apart Values

Destructuring Structs
Destructuring Enums
Destructuring Nested Structs and Enums
Destructuring Structs and Tuples

Ignoring Values in a Pattern

Ignoring an Entire Value with _
Ignoring Parts of a Value with a Nested _
Ignoring an Unused Variable by Starting Its Name with _
Ignoring Remaining Parts of a Value with .. : using .. must be unambiguous. If it is unclear which values are intended for matching and which should be ignored, Rust will give us an error.

Extra Conditionals with Match Guards

A match guard is an additional if condition specified after the pattern in a match arm that must also match, along with the pattern matching, for that arm to be chosen. Match guards are useful for expressing more complex ideas than a pattern alone allows.

@ Bindings

The at operator (@) lets us create a variable that holds a value at the same time we’re testing that value to see whether it matches a pattern.
Using @ lets us test a value and save it in a variable within one pattern.

Summary

Rust’s patterns are very useful in that they help distinguish between different kinds of data. When used in match expressions, Rust ensures your patterns cover every possible value, or your program won’t compile. Patterns in let statements and function parameters make those constructs more useful, enabling the destructuring of values into smaller parts at the same time as assigning to variables. We can create simple or complex patterns to suit our needs.