19.2. Advanced Traits
Specifying Placeholder Types in Trait Definitions with Associated Types

Associated types connect a type placeholder with a trait such that the trait method definitions can use these placeholder types in their signatures. The implementor of a trait will specify the concrete type to be used in this type’s place for the particular implementation. That way, we can define a trait that uses some types without needing to know exactly what those types are until the trait is implemented.

```rust
impl Iterator for Counter {
    type Item = u32;

    fn next(&mut self) -> Option<Self::Item> {
        // --snip--
```

Default Generic Type Parameters and Operator Overloading

Fully Qualified Syntax for Disambiguation: Calling Methods with the Same Name

Using Supertraits to Require One Trait’s Functionality Within Another Trait

Using the Newtype Pattern to Implement External Traits on External Types

Creating a new type in a tuple struct.The tuple struct will have one field and be a thin wrapper around the type we want to implement a trait for. Then the wrapper type is local to our crate, and we can implement the trait on the wrapper.